+++ taskmanager-web/src/main/web/src/app/app.module.ts
@@ -1,8 +1,8 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { AppComponent } from './app.component';
-import { AddTaskModule } from './addtask/addtask.module';
+import { AppComponent } from './app.component'; 
+import { UpdateTaskModule } from './updatetask/updatetask.module'; 
 import { ViewTaskModule } from './viewtask/viewtask.module';
 import { RequestOptions, RequestMethod, Headers } from '@angular/http';
 import { AppRoutingModule } from './app-routing.module'; 
@@ -28,7 +28,7 @@ export class MyOptions extends RequestOptions {
     AppComponent
   ],
   imports: [
-    BrowserModule, FormsModule, CommonModule, AppRoutingModule, AddTaskModule, ViewTaskModule, HttpModule, HttpClientModule, NgbModule.forRoot(), NgbTypeaheadModule.forRoot()
+    BrowserModule, FormsModule, CommonModule, AppRoutingModule, UpdateTaskModule, ViewTaskModule, HttpModule, HttpClientModule, NgbModule.forRoot(), NgbTypeaheadModule.forRoot()
   ],
   exports: [
     FormsModule
