+++ taskmanager-web/src/main/web/src/app/updatetask/updatetask.component.html
@@ -1,7 +1,7 @@
 <div class="row my-3 p-3 bg-white rounded box-shadow">
     <div class="col-xs-12 col-sm-2 col-md-2"></div>
     <div class="col-xs-12 col-sm-8 col-md-8">
-            <form #addTaskForm="ngForm" method="post" (ngSubmit)="addTask(task)" novalidate >
+            <form #updateTaskForm="ngForm" method="post" (ngSubmit)="updateTask(task)" novalidate >
                 <div class="form-group row">
                     <label for="taskName" class="col-sm-2 col-form-label">Task Name</label>
                     <div class="col-sm-10">
@@ -58,8 +58,14 @@
                                 </div>
                             </div>
                         <div class="input-group">
-                            <button type="submit" class="btn btn-success mx-3" [disabled]="(task.taskName === '' || fromDate === null || toDate === null)">Add Task</button>
-                            <button type="button" class="btn btn-info mx-5" (click)=resetButton()>Reset</button>
+                            <div *ngIf="flow === 'addtask'">
+                                    <button type="submit" class="btn btn-success mx-3" [disabled]="(task.taskName === '' || fromDate === null || toDate === null)">Add Task</button>
+                                    <button type="button" class="btn btn-info mx-5" (click)=resetButton()>Reset</button>
+                            </div>
+                            <div *ngIf="flow === 'updatetask'">
+                                    <button type="submit" class="btn btn-success mx-3" [disabled]="(task.taskName === '' || fromDate === null || toDate === null)">Update</button>
+                                    <button type="button" class="btn btn-info mx-5" (click)=viewTaskScreen()>Cancel</button>
+                            </div>
                         </div>
                     </div>
                 </div>
@@ -86,8 +92,8 @@
         </div>
         <!-- Modal footer -->
         <div class="modal-footer">
-          <button type="button" class="btn btn-danger" data-dismiss="modal" (click)=resetButton()>Back To Add Task</button>
-          <button type="button" class="btn btn-success" (click)=viewTaskScreen()>Navigate to View Task</button>
+          <button type="button" class="btn btn-danger" *ngIf="flow === 'addtask'" data-dismiss="modal" (click)=resetButton()>Back To Add Task</button>
+          <button type="button" class="btn btn-success" data-dismiss="modal" (click)=viewTaskScreen()>Navigate to View Task</button>
         </div>
       </div>
     </div>
